Make loader.js usable as a script
Despite scheduling the call to `require` with `process.nextTick` or
`setTimeout`, I was frequently encounting an error from the file being
loaded before the download had completed. As a workaround, make it so
that loader.js can be run as a stand-alone script to download the
bundles, as a separate step before starting the server.